requirements.txt
# Flask 核心框架
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5

# 数据库
PyMySQL==1.1.0

# 加密和安全
cryptography==41.0.7
python-dotenv==1.0.1
Werkzeug==3.0.1

# HTTP 请求
requests==2.31.0

# WebSocket 和异步
gevent==23.9.1
gevent-websocket==0.10.1

# 工具库
python-dateutil==2.8.2
six==1.16.0

# 开发工具
gunicorn==21.2.0
